# mvp-usluge/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MVP Usluge API",
    "version": "1.0.0",
    "description": "REST API dokumentacija za platformu oglašavanja uslužnih aktivnosti",
    "contact": {
      "name": "API Support",
      "email": "support@mvp-usluge.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://mvp-usluge.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "NextAuth.js JWT token"
      }
    },
...
  },
...
  "paths": {
...
    "/api/cron/verify-companies": {
      "get": {
        "summary": "Proverava i ažurira verifikaciju preduzeća",
        "description": "Cron job - proverava da li su preduzeća verifikovana",
        "tags": [
          "Cron"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Verifikacija obavljena"
          },
          "401": {
            "description": "Unauthorized (zahteva CRON_SECRET)"
          }
        }
      },
      "post": {
        "summary": "Proverava i ažurira verifikaciju preduzeća",
        "description": "Cron job - proverava da li su preduzeća verifikovana",
        "tags": [
          "Cron"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Verifikacija obavljena"
          },
          "401": {
            "description": "Unauthorized (zahteva CRON_SECRET)"
          }
        }
      }
    },
...
  }
}